hw5

+++ source/Engine/Rect.hpp
@@ -50,27 +50,27 @@ namespace Math
         Math::ivec2 point_1{ 0, 0 };
         Math::ivec2 point_2{ 0, 0 };
 
-        double Left() const noexcept
+        int Left() const noexcept
         {
             return std::min(point_1.x, point_2.x);
         }
 
-        double Right() const noexcept
+        int Right() const noexcept
         {
             return std::max(point_1.x, point_2.x);
         }
 
-        double Top() const noexcept
+        int Top() const noexcept
         {
             return std::max(point_1.y, point_2.y);
         }
 
-        double Bottom() const noexcept
+        int Bottom() const noexcept
         {
             return std::min(point_1.y, point_2.y);
         }
 
-        Math::vec2 Size() const noexcept
+        Math::ivec2 Size() const noexcept
         {
             return { Right() - Left(), Top() - Bottom() };
         }
